Impl. dialogs in desktop
Desktop is now able to inject a workitem as a dialog
in addition of the current workitem.

I use a search param did= for injecting a dialogId into the
main workitem view router.

After that, views can impl. presence of a dialog on the current
workitem.

A dialog is a workitem, but it's not stated as current.

add-workitem quest, add a dialog if the wi. kind is set to "dialog".
you must call remove-dialog quest for navigating out and releasing
the added dialog.

!!!this is a WIP behavior!!!