Refactor workitem-builder and impl. main-button
an after-new quest is now triggered for loading
the entity graph, and triggering optional onLoad and mainButton
hooks / handlers.

hooks names are now normalized, and ready to handle a cancelation token
for the day when...

a mainButton impl. is now available for overriding the publish/submit
button in the editor.

a published entity is now only submited when the main button is actioned.