WIP: experiment react-tooltip (fix tooltip rebuilding and use CSS)
Note that the CSS can be used only in the div content and not for the
border. it's a bit difficult to change the border style when it's
necessary to support the triangles trick based on the border styles.

The react component can be used everywhere. It's just a problem when the
tooltip is visible and the switch is done on the configurator.