Kill the wizard from the desktop (the creator) RAII
Use the done event for that. But note that it's not right because the
wizard must be killed only when it's finished. ANd the done event is
sent by a quest qhich can be run in parallel of other quests of this
wizard.

IMO, the best way will be to use the new serie scheduler queue for a new
last quest for the wizard. And this quest could be started only if /
when the parallel queue is empty. Of course, no other quest can be
started when a quest is running in the serie queue. It's still not fully
implemented in the scheduler.