Fix navigation bug when maxInstances is reached
The navigation behavior of maxInstances is:

When the user is in another workitem, and he click "search" (limited workitem)
we simply navigate to the existing instance insteed of creating a new one.

When a max instance is reached,
if no current search in location is available, it's mean that the existing workitem
is not open currently in the UI when the user as clicked (a second time, too early)
so we don't need to navigate,

simply skipping.