Hinter UX: Tab key don't valide current hinter row  anymore, user must use Enter ok mouse click.
Tab is usefull for navigating to the next field, this ensure not adding something
when leaving hinter row.

Hinter don't hide after a validation, this permit adding multiple
choice in case of hinter+plugin, avoiding loose of hinter context.

I can perhaps restore the Tab for validating, when hinter is not used in plugin.